StandardLibrary/runtimes/python/pyproject.toml
[tool.poetry]
name = "standard-library"
version = "0.1.0"
description = ""
authors = ["AWS <aws-arg-dafny@amazon.com>"]
packages = [
    { include = "standard_library", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11.0"
pytz = "^2023.3.post1"
standard-library-externs = { path = "src/standard_library/extern", develop = false }

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
standard-library-test-internaldafny = { path = "test/internal_generated_dafny", develop = false}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
